requirements.txt
pykrx>=1.0.45
anthropic>=0.42.0
requests>=2.31.0
pandas>=2.1.0
python-dotenv>=1.0.0
schedule>=1.2.0
streamlit>=1.40.0
plotly>=5.18.0
openpyxl>=3.1.2
beautifulsoup4>=4.12.0
lxml>=5.1.0
